+++ src/main/java/zb/accountMangement/common/service/RedisService.java
@@ -1,4 +1,4 @@
-package zb.accountMangement.common.util;
+package zb.accountMangement.common.service;
 
 
 import java.util.concurrent.TimeUnit;
@@ -12,7 +12,7 @@ import zb.accountMangement.member.dto.SmsVerificationDto;
 
 @Service
 @RequiredArgsConstructor
-public class RedisUtil {
+public class RedisService {
   private final RedisTemplate<String, Object> redisTemplate;
   private final StringRedisTemplate stringRedisTemplate;
